feat(): Affichage des murs de droite et de gauche.
